# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transparência API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/predios": {
            "get": {
                "tags": [
                    "/predios"
                ],
                "summary": "listar predios",
                "operationId": "456fe57437d1670dc58d8b8a79b12d8c",
                "responses": {
                    "200": {
                        "description": "sucesso"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/predios/bloco/{bloco}": {
            "post": {
                "tags": [
                    "/predios"
                ],
                "summary": "Cadastrar um predios numa bloco",
                "operationId": "3e2cbe2ec69551301de9021052b49a81",
                "parameters": [
                    {
                        "name": "bloco",
                        "in": "path",
                        "description": "id da bloco",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "c_descpredi": {
                                        "description": "denominação do Prédio",
                                        "type": "string"
                                    },
                                    "c_entrpredi": {
                                        "description": "Entrada do prédio",
                                        "type": "string"
                                    },
                                    "n_codicoord": {
                                        "description": "id do Coordenador do prédio",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "bloco cadastrado com sucesso"
                    },
                    "412": {
                        "description": "Erro ao validar os dados"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/predios/{predio}": {
            "get": {
                "tags": [
                    "/predios"
                ],
                "summary": "mostrar um predio",
                "operationId": "db6cd45c9e5711d9d73696ef2dd1a0bd",
                "parameters": [
                    {
                        "name": "predio",
                        "in": "path",
                        "description": "id do predio",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sucesso"
                    },
                    "404": {
                        "description": "predio não encontrado"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "/predios"
                ],
                "summary": "atualizar os dados de um predios",
                "operationId": "6fdcb1219d2e18883ad2a1e7c1059ecb",
                "parameters": [
                    {
                        "name": "predio",
                        "in": "path",
                        "description": "id da bloco",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "c_descpredi": {
                                        "description": "denominação do Prédio",
                                        "type": "string"
                                    },
                                    "c_entrpredi": {
                                        "description": "Entrada do prédio",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "predio Atualizado com sucesso"
                    },
                    "412": {
                        "description": "Erro ao validar os dados"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "/predios"
                ],
                "summary": "apagar um predio",
                "operationId": "f64f25d3fb7795f4571b69c4e79a1003",
                "parameters": [
                    {
                        "name": "predio",
                        "in": "path",
                        "description": "id do predio",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "predio deletado com sucesso!"
                    },
                    "404": {
                        "description": "predio não encontrado"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}